Added Azure DevOps CI/CD (#442)
* move changes from another branch

* revert redutent changes

* add minor changes

* changes of jenkins

* change bundle install position

* test/cucumber: ignore dylib injection when not on Test Cloud

* Include appcenter.sh to azure pipelines yaml (#443)

* update workspace

* fix appcenter.sh

* disable xctest

* debug gemrc

* write gemrc

* add -e

* revert test changes

* enable xctests

* some minor changes

* changes cucumber rb phines

* chnage versions

* Fix logic to switch to the second tab (#444)

* enable test + add while

* move code to the function

* Revert "Fix logic to switch to the second tab (#444)" (#445)

This reverts commit 4d2a6aea53e4a208e9a803beb97f7db96eaea1cc.

* add condition

* Update azure-pipelines.yml

* Update azure-pipelines.yml

try to add Xcode 10.3

* Fix logic to switch to the second tab (#446)

* enable test + add while

* move code to the function

* fix attempt_count-1

* 5 -> 20 retries

* increase retry numbers

* play with attempt number

* change back to infinity loop

* remove `attempt_count`

* V malob/ci test v3 (#1)

* resolve comments

* remove steps for testing

* remove beta condition

* add 1 second sleep

* sleep(5)

* 100 attempts

* 0.5 * 10000

* try real values

* remove debug info

* remove 10.3 for now

* include azurepipelines to CI condition

* add comment to describe `mainJob` variable

* remove legacy comment and travis condition

+++ bin/make/app-cal.sh
@@ -8,9 +8,11 @@ ensure_valid_core_sim_service
 # Command line builds alway make a fresh framework
 banner "Ensure the calabash.framework"
 
-FRAMEWORK="calabash.framework"
-rm -rf "${FRAMEWORK}"
-make framework
+if [ "${MAKE_FRAMEWORK}" != "0" ]; then
+  FRAMEWORK="calabash.framework"
+  rm -rf "${FRAMEWORK}"
+  make framework
+fi
 
 banner "Preparing to build LPTestTarget"
